% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Calculates the likelihood for each data point and each latent class}
\usage{
likelihood(vectorized_probs, responses)
}
\arguments{
\item{vectorized_probs}{A list containing:
\itemize{
\item vecprobs: vector of outcome probabilities, a flattened list of matrices
\itemize{
\item dim 0: for each outcome
\item dim 1: for each category
\item dim 2: for each cluster
\item in other words, imagine a nested loop, from outer to inner:
\itemize{
\item for each cluster, for each category, for each outcome
}
}
\item numChoices: vector, number of outcomes for each category
\item classes: integer, number of classes (or clusters)
Can be the output of poLCAParallel.vectorize()
}}

\item{responses}{A design matrix of responses}
}
\value{
A matrix of likelihoodswith dimensions n_data x n_cluster
}
\description{
Calculates the likelihood for each data point and each latent class
}
