% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCAParallel.goodnessfit.R
\name{poLCAParallel.goodnessfit}
\alias{poLCAParallel.goodnessfit}
\title{Add predcell, Gsq and Chisq to a fitted poLCA object}
\usage{
poLCAParallel.goodnessfit(results)
}
\arguments{
\item{results}{A model object estimated using the \code{poLCA} function}
}
\value{
The model object with three added items:
\itemize{
\item \code{predcell}: dataframe of unique responses with their observed and expected
frequencies
\item \code{Gsq}: log likelihood ratio
\item \code{Chisq}: chi squared statistic
}
}
\description{
Finds unique responses and put them in a dataframe along with the observed
and expected frequencies. Also calculates the log likelihood ratio and chi
squared statistics.
}
\details{
Wrapper function around the C++ function \code{GoodnessFitRcpp()}.
}
