% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{poLCAParallel}
\alias{poLCAParallel}
\title{poLCAParallel: Parallelized Polytomous Variable Latent Class Analysis}
\description{
The package poLCAParallel reimplements poLCA fitting, standard error
calculations, goodness of fit tests and the bootstrap log-likelihood ratio
test in C++.
}
