% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCAParallel.unvectorize.R
\name{poLCAParallel.unvectorize}
\alias{poLCAParallel.unvectorize}
\title{Reverse the operations of poLCAParallel.vectorize()}
\usage{
poLCAParallel.unvectorize(vp)
}
\arguments{
\item{vp}{list of three items (vecprobs, numChoices, classes) where
* vecprobs: vector of outcome probabilities, a flattened list of matrices
   * dim 0: for each outcome
   * dim 1: for each category
   * dim 2: for each cluster
   * in other words, imagine a nested loop, from outer to inner:
      * for each cluster, for each category, for each outcome
* numChoices: integer vector, number of outcomes for each category
* classes: integer, number of classes (or clusters)}
}
\value{
list of length n_category. For the ith entry, it contains a
matrix of outcome probabilities with dimensions n_class x n_outcomes[i]
}
\description{
Mimics poLCA.unvectorize but with some of the dimensions swapped. Given the
return value, or even modified, of poLCAParallel.vectorize(), return a list
of matrices containing outcome probabilities.
}
