% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCAParallel.vectorize.R
\name{poLCAParallel.vectorize}
\alias{poLCAParallel.vectorize}
\title{Put all outcome probabilities into a single vector}
\usage{
poLCAParallel.vectorize(probs)
}
\arguments{
\item{probs}{list of length n_category. For the ith entry, it contains a
matrix of outcome probabilities with dimensions n_class x n_outcomes[i]}
}
\value{
a list containing:
  * vecprobs: vector of outcome probabilities, a flattened list of matrices
    * dim 0: for each outcome
    * dim 1: for each category
    * dim 2: for each cluster
    * in other words, imagine a nested loop, from outer to inner:
        * for each cluster, for each category, for each outcome
  * numChoices: vector, number of outcomes for each category
  * classes: integer, number of classes (or clusters)
}
\description{
Given a list of matrices of outcome probabilities, flatten it into a vector,
suitable for EmAlgorithmRcpp. It mimics poLCA.vectorize() but with some of
the dimensions swapped to improve cache efficiency in the C++ code.
}
