% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrt.R
\name{blrt}
\alias{blrt}
\title{Bootstrap likelihood ratio test (no regression only)}
\usage{
blrt(
  model_null,
  model_alt,
  n_bootstrap,
  n_thread = parallel::detectCores(),
  n_rep = 1,
  max_iter = 1000,
  tol = 1e-10
)
}
\arguments{
\item{model_null}{Fitted poLCA object, the null model}

\item{model_alt}{Fitted poLCA object, the alt model}

\item{n_bootstrap}{Number of bootstrap samples}

\item{n_thread}{Number of threads}

\item{n_rep}{Number of initial values to try when fitting on the bootstrap
samples}

\item{max_iter}{Maximum number of iterations for EM algorithm}

\item{tol}{Tolerance for difference in log likelihood, used for
stopping condition}
}
\value{
List containing the following:
<ul>
  <li>fitted_log_ratio: log likelihood ratio comparing the null and alt
model</li>
  <li>bootstrap_log_ratio: vector of length n_bootstrap, bootstrapped log
likelihood ratio comparing the null and alt model</li>
  <li>p_value<: the porportion of bootstrap samples with log likelihood
ratios greater than the fitted log likelihood ratio/li>
</ul>
}
\description{
Does the bootstrap likelihood ratio test. Provide two fitted models, the
null model and the alt model which fits a different number of clusters.
Bootstrap samples are generated using the null model. The null model and alt
model are refitted on the bootstrap samples to investigate the log likelihood
ratio of the two models.
}
\details{
Runs in parallel for each bootstrap sample, potentially high memory if the
data is large
}
