% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCAParallel.goodnessfit.R
\name{poLCAParallel.goodnessfit}
\alias{poLCAParallel.goodnessfit}
\title{Add predcell, Gsq and Chisq to a fitted poLCA object (or a list)}
\usage{
poLCAParallel.goodnessfit(results)
}
\arguments{
\item{results}{a poLCA object (see poLCA.R)}
}
\value{
list with 3 items:
  * predcell: dataframe of unique responses with their observed and expected
    frequencies
  * Gsq: log likelihood ratio
  * Chisq: chi squared statistic
}
\description{
Wrapper function around the C++ function GoodnessFitRcpp(). Finds unique
responses and put them in a dataframe along with the observed and
expected frequencies. Also calculates the log likelihood ratio and chi
squared statistics.
}
