% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCAParallel.se.R
\name{poLCAParallel.se}
\alias{poLCAParallel.se}
\title{Calculate the standard errors and adds them to the poLCA object}
\usage{
poLCAParallel.se(polca, is_smooth = FALSE)
}
\arguments{
\item{polca}{the resulting poLCA object from calling poLCA()}

\item{is_smooth}{Logical, experimental, for calculating the standard errors,
whether to smooth the outcome probabilities to produce more numerical
stable results as a cost of bias.}
}
\value{
the poLCA object with the attributes $P.se, $probs.se, $coeff.se and
$coeff.V modified or added
}
\description{
R wrapper function for the C++ function StandardError Rcpp
}
\details{
Calculate the standard errors and adds them, as attributes, to the poLCA
object, $P.se, $probs.se $coeff.se and $coeff.V

$coeff.se and $coeff.V are set to NA if the provided poLCA is a standard
poLCA with no regression
}
